Add a clang-format config file.
The rewrites done to our code aren't ... entirely great, and I don't have the
time to walk over each file and individually fix any clang-format mistakes or
awkwardmess.

I'm adding the config file so that we can fix issues over time, and any file
that I go to edit will get converted into our New One True Coding Style.  I
don't mean for this to be a pedantic coding style, and if your change looks
more readable with a line width of 105 characters, then go for it.

Note that CommentPragmas means //TraceEvent lines are ignored and not comment
reflowed, and any clang-format weirdness on a line can either be disabled
via...

    // clang-format: off
        The code to ignore
    That has some weird formatting
        Intentionally
    // clang-format: on

Or for a single line

    ar & a & b & c;  // clang-format: ignore

This is also not intended to be the inarguable final form of our coding style
config.  It's totally possible that I've missed something, or that a different
line length would end up causing our existing code to look better.