Enabled /etc/foundationdb portability
We store a backup of fdb.cluster and foundationdb.conf to /etc/foundationdb.default during build, and copy them to /etc/foundationdb during startup if those files do not exist.

This enables the use of a Docker volume at /etc/foundationdb with or without configuration files, so we can both automate pulling of default files and/or use custom files. This implementation also supports mounting a volume with only one of two files, should we want to use a custom configuration for one and leave default for the other one.