Optimize truncate calls
A call to truncate currently iterates through the
whole page list. This is very inefficient, especially
if truncate only affects a small portion of the file.

This fix will take into account truncate calls that
make the file larger (in which case it simply skips
the cache invalidation part) or calls that make a file
smaller only by a few pages. In the later case, we
look up all affected pages in the hash table.