WiP:Has bug: assign roles and distribute files
The global variables for a process in real mode are separate across processes in real mode, BUT
they are NOT separate in simulation mode.
The workers in different processes can still access the global variables,
which makes one worker overwrite another work's global variable value.

To solve this problem, which happens in other FDB code as well,
we allocate the global variable in a heap and pass the pointers across functions,
as what DataDistribution.actor does.

The other approach is to create two code path: one for real mode and the other for simulator,
as the g_simulator does to create the process context for each simulated process