Fix race condition in waiting for watch on trigger keys
We use watch to let processes know if the restore request is ready and if the restore has been finished.
When we setup the watch and wait on the watch on a key, we ASSUME that the key has not been set yet.
However, under certain situations (e.g., the restore is fast and the restore request agent is slow), the trigger key may have been set before we wait on the watch.
Without handling this situation, the system will stuck in waiting on the watch.

To solve this situation, we need to check the existance of the key (the watch is on) before we wait on the watch