Skip log block that does not have all mutations in the kv pair
The last kv pair in the log data block may need to be concatenated with the next log data block.
When this happens, we skip the kv pair for sampling data.
We cannot skip it for the real restoring workload.
If we skipped the kv for restoring workload, we will get inconsistent database after restoring

We also avoid printing the parse error to stderr to avoid false positive in correctness test