let master wait for any applier reply at apply db
Applier may crash in applying mutations.
Node crash may make master waits infinitely for the reply from all nodes.

Change waitForAll semantics to waitForAny when waiting for the appliers response for applying mutations to DB

This is a workaround. The long-term solution should handle the failure in a better way