TeamRemover: No order of removing team and machine team
We do NOT enforce the removing order of removing a machine team
and the server teams on the machine team.

This is for the benefit of clear code logic.

When a storage server locality changes, we first remove the server
and its machine if needed, before we handle the server team removal
and addition.