Force Make to push $CC into the environment.
This fixes a long standing weird issue of doing non-docker builds of FDB
can hit problems of $CC not being defined in link-wrapper.sh.  It turns
out that this is because the official docker image defines CC in the
environment, and no one else does that.

Instead, we can just force Make to propagate its setting of CC into the
environment, thus requiring no extra configuration to be able to do a
build.